Adds author view
